+++ src/routes/_authenticated/grading.$id.tsx
@@ -1,12 +1,12 @@
-import { createFileRoute } from '@tanstack/react-router'
+import { createFileRoute } from "@tanstack/react-router"
+import EditGradingForm from "@/features/grading/forms/edit-grading-form"
 
-export const Route = createFileRoute('/_authenticated/grading/$id')({
+export const Route = createFileRoute("/_authenticated/grading/$id")({
   component: RouteComponent,
 })
 
 function RouteComponent() {
-    const { id } = Route.useParams()
-  return (<div>
-      <div className='text-2xl'>Incoming ID: {id}</div>
-  </div>)
+  const { id } = Route.useParams()
+
+  return <EditGradingForm gatePassId={id} />
 }